SelectPickerItemsをPlatform毎に分けないようにしました
