+++ example-app/SantokuApp/src/components/picker/SelectPickerItems.tsx
@@ -1,19 +1,32 @@
-import {PickerProps} from '@react-native-picker/picker';
-import {log} from 'framework/logging';
+import {Picker, PickerProps} from '@react-native-picker/picker';
 import React from 'react';
-import {Platform, ViewProps} from 'react-native';
 
 import {Item} from './SelectPicker';
 
 export type SelectPickerItemsProps<ItemT> = PickerProps<React.Key> & {
   items: Item<ItemT>[];
   keyExtractor?: ((item: ItemT, index: number) => React.Key) | undefined;
-  pickerItemsContainerProps?: ViewProps;
 };
 
-export const SelectPickerItems = <ItemT extends unknown>(props: SelectPickerItemsProps<ItemT>) => {
-  React.useEffect(() => {
-    log.warn(`SelectPickerItems is not supported on: ${Platform.OS}`);
-  }, []);
-  return null;
+export const SelectPickerItems = <ItemT extends unknown>({
+  selectedValue,
+  onValueChange,
+  items,
+  keyExtractor,
+  ...rest
+}: SelectPickerItemsProps<ItemT>) => {
+  return (
+    <Picker selectedValue={selectedValue} onValueChange={onValueChange} {...rest}>
+      {items.map((item, index) => (
+        <Picker.Item
+          key={keyExtractor ? keyExtractor(item.value, index) : item.key ?? item.label}
+          label={item.label}
+          value={item.key ?? item.value}
+          color={item.color}
+          fontFamily={item.fontFamily}
+          style={item.style}
+        />
+      ))}
+    </Picker>
+  );
 };
